+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 - Added `Client#create_candidate`
 - Added `Client#create_job_application`
 
+## v0.3.3 - 2021-03-10
+
+- Added `Client#create_candidate`
+- Added `Client#create_job_application`
+- Fix accessing `Client#jobs` without filters
+
 ## v0.3.2 - 2021-02-03
 
 - Add support for `CustomField::Select` and `CustomField::MultiSelect`